Address review: remove DCL, split types into separate files
- Remove outer null check in InitializeCore (single-threaded init via
  startup hook, lock alone is sufficient)
- Split ITypeMapWithAliasing.cs into three files: ITypeMapWithAliasing.cs,
  SingleUniverseTypeMap.cs, AggregateTypeMap.cs
